.make#{,un}install: not uninstalling some files
Writing to .installed in xargs meant there were data races, leading to many files being skipped. This commit reworks writing to .installed with some duplicate code. We still don't directly find and uninstall, as changing the name of or removing a file in the repository would lead to not being able to keep track of that file if it was already installed.